Big update: redesign of the MCV system, redesign of namespace, organization of files within the application, Creation of the contact page, corrections of several bugs

+++ view/backend/adminHome.php
@@ -4,47 +4,50 @@ ob_start();
 ?>
 <h2>Salut Jean, comment ça va ?</h2>
 <h4>Que puis-je vous proposer aujourd'hui ?</h4>
-<h5>Continuer sur un chapitre commencé ?</h5>
 
 <?php
-
-while ($data = $posts->fetch()) {
-    if ($data['blogPostStatus'] == 'inProgress') {
-        $tricolore = 'Red';
-
-?>
-        <div class="p-2 m-2 rounded bg-info d-flex justify-content-between ">
-            <div class="">
-                <?= $data['blogPostId'] ?> • <?= $data['blogPostTitle'] ?><br>
-                <?= substr($data['blogPostContents'], 0, 100) ?>
-            </div>
-            <div class="text-center">
-                <a href="index.php?action=modifyPost&id=<?= $data['blogPostId'] ?>" class="text-dark" title="Pour corriger ce chapitre"><i class="p-2 fas fa-pencil-alt fa-lg "></i></a>
-                <a href="index.php?action=modifyPostStatus&id=<?= $data['blogPostId'] ?>" title="Pour ouvrir ce chapitre à la lecture">
-                    <img src="public/image/tricolor<?= $tricolore ?>.png" alt="Feux tricolores indiquant le status de l'article">
-                </a>
-
-                <a href="index.php?action=deletePost&id=<?= $data['blogPostId'] ?>" class="text-dark" title="Pour supprimer ce chapitre"><i class="p-2 fas fa-trash-alt fa-lg"></i></a>
-
-            </div>
-
-        </div>
-<?php
+foreach($posts as $data) {
+    if ($data['blogPostStatus'] == 'inProgress') 
+    {
+        echo "<h5>Continuer de travailler sur un chapitre commencé ?</h5>";
+    break;
     }
 }
-$result = $comments->fetchAll();
+
+foreach($posts as $data) {
+    if ($data['blogPostStatus'] == 'inProgress') 
+    {
+        $tricolore = 'Red';
+    ?>
+    
+    <div class="p-2 m-2 rounded bg-info d-flex justify-content-between ">
+        <div class="">
+            <?= $data['blogPostId'] ?> • <?= $data['blogPostTitle'] ?><br>
+            <?= substr($data['blogPostContents'], 0, 100) ?>
+        </div>
+        <div class="text-center">
+            <a href="index.php?action=modifyPost&id=<?= $data['blogPostId'] ?>" class="text-dark" title="Pour corriger ce chapitre"><i class="p-2 fas fa-pencil-alt fa-lg "></i></a>
+            <a href="index.php?action=modifyPostStatus&id=<?= $data['blogPostId'] ?>" title="Pour ouvrir ce chapitre à la lecture">
+                <img src="public/image/tricolor<?= $tricolore ?>.png" alt="Feux tricolores indiquant le status de l'article">
+            </a>
+            <a href="index.php?action=deletePost&id=<?= $data['blogPostId'] ?>" class="text-dark" title="Pour supprimer ce chapitre"><i class="p-2 fas fa-trash-alt fa-lg"></i></a>
+        </div>
+    </div>
+    <?php
+    }
+}
 
 ?>
 
 
 <?php
 $countDanger = 0;
-foreach ($result as $dataComment){
-
-    if ($dataComment['commentReport'] == 'reported'  && $dataComment['startingCommentId'] == 0) {
+foreach ($comments as $dataComment){
+    if ($dataComment['commentReport'] == 'reported'  && $dataComment['startingCommentId'] == 0) 
+    {
         $countDanger++;
         if ($countDanger == 1) {echo"<h5>Il y a de nouveaux commentaires signalés :</h5>";}
-?>
+        ?>
         <div class="d-flex flex-row mb-4 border border-danger">
             <div class="d-flex flex-column col-11 sansPadding ">
                 <div class="d-flex flex-row bg-danger text-white">
@@ -56,17 +59,13 @@ foreach ($result as $dataComment){
                     <div class="commentsNews col"><?= $dataComment['commentDate_fr']; ?></div>
                 </div>
                 <div>
-                    <?php
-
+                <?php
                     echo '<p class="commentaireTexte">' . htmlspecialchars($dataComment['commentContents']) . '</p>';
-
-                    if ($data['answerId'] != 0) {
+                    if ($dataComment['answerId'] != 0) {
                         echo '<hr><p class="commentaireTexte">Votre réponse : <br>';
                         echo '<em>' . $dataComment['answerContents'] . '</em></p>';
                     }
-                    ?>
-
-
+                ?>
                 </div>
             </div>
             <div class="p-2 d-flex col-1 text-center pictoCommentsAdmin">
@@ -75,19 +74,20 @@ foreach ($result as $dataComment){
                 <a href="index.php?action=commentsDelete&id=<?= $dataComment['commentId'] ?>" class="text-dark" title="Pour supprimer le commentaire"><i class="p-2 fas fa-trash-alt fa-lg"></i></a>
             </div>
         </div>
-<?php
+    <?php
     }
     unset($dataComment);
 }
 ?>
 <?php
-$countWarning =0;
-foreach ($result as $dataComment){
-
-    if ($dataComment['commentReport'] == 'waiting' && $dataComment['startingCommentId'] == 0) {
+$countWarning = 0;
+foreach ($comments as $dataComment)
+{
+    if ($dataComment['commentReport'] == 'waiting' && $dataComment['startingCommentId'] == 0) 
+    {
         $countWarning++;
-        if ($countDanger == 1) {echo"<h5>Il y a de nouveaux commentaires que vous n'avez pas validé :</h5>";}
-?>
+        if ($countWarning == 1) {echo"<h5>Il y a de nouveaux commentaires que vous n'avez pas validé :</h5>";}
+        ?>
         <div class="d-flex flex-row mb-4 border border-warning">
             <div class="d-flex flex-column col-11 sansPadding ">
                 <div class="d-flex flex-row bg-warning text-white">
@@ -99,34 +99,29 @@ foreach ($result as $dataComment){
                     <div class="commentsNews col"><?= $dataComment['commentDate_fr']; ?></div>
                 </div>
                 <div>
-                    <?php
-
-                    echo '<p class="commentaireTexte">' . htmlspecialchars($dataComment['commentContents']) . '</p>';
-
-                    if ($data['answerId'] != 0) {
-                        echo '<hr><p class="commentaireTexte">Votre réponse : <br>';
-                        echo '<em>' . $dataComment['answerContents'] . '</em></p>';
-                    }
-                    ?>
-
-
+                <?php
+                echo '<p class="commentaireTexte">' . htmlspecialchars($dataComment['commentContents']) . '</p>';
+                if ($dataComment['answerId'] != 0) 
+                {
+                    echo '<hr><p class="commentaireTexte">Votre réponse : <br>';
+                    echo '<em>' . $dataComment['answerContents'] . '</em></p>';
+                }
+                ?>
                 </div>
             </div>
             <div class="p-2 d-flex col-1 text-center pictoCommentsAdmin">
                 <a href="index.php?action=commentsValidation&id=<?= $dataComment['commentId'] ?>" class="text-dark" title="Pour valider le commentaire"><i class="p-2 fas fa-check-square fa-lg"></i></a>
                 <a href="index.php?action=commentsModify&id=<?= $dataComment['commentId'] ?>" class="text-dark" title="Pour répondre au commentaire ou corriger votre réponse"><i class="p-2 fas fa-pencil-alt fa-lg "></i></a>
-                <a href="index.php?action=commentsDelete&id=<?= $dataComment['commentId'] ?>" class="text-dark" title="Pour supprimer le commentaire"><i class="p-2 fas fa-trash-alt fa-lg"></i></a>
+                <a href="index.php?action=commentsDelete&id=<?= $dataComment['commentId'] ?>" onclick=" return confirm('Etes-vous sûr de vouloir supprimer ce commentaire ? ');" class="text-dark" title="Pour supprimer le commentaire"><i class="p-2 fas fa-trash-alt fa-lg"></i></a>
             </div>
         </div>
-<?php
+    <?php
     }
     unset($dataComment);
 }
 
-$comments->closeCursor();
-$posts->closeCursor();
 
 $content = ob_get_clean();
-require('view/template.php');
+require('view/frontend/template.php');
 
 ?>
