+++ view/backend/commentsViewAdmin.php
@@ -3,7 +3,7 @@
 ob_start(); 
 $chap = 1;
 echo "<h2>Chap. ". $chap ."</h2>";
-while ($data = $comments->fetch())
+foreach($comments as $data)
     {
         if ($data['commentBlogPostId'] != $chap && $data['startingCommentId'] == 0)
         {
@@ -50,19 +50,19 @@ while ($data = $comments->fetch())
                 </div>
             </div>
             <div class="p-2 d-flex col-1 text-center pictoCommentsAdmin">
+                <?php $message="'Etes-vous sûr de vouloir supprimer cette entrée?'"; ?>
                 <a href="index.php?action=commentsValidation&id=<?= $data['commentId'] ?>" class="text-dark" title="Pour valider le commentaire"><i class="p-2 fas fa-check-square fa-lg"></i></a>
                 <a href="index.php?action=commentsModify&id=<?= $data['commentId'] ?>" class="text-dark" title="Pour répondre au commentaire ou corriger votre réponse"><i class="p-2 fas fa-pencil-alt fa-lg "></i></a>
-                <a href="index.php?action=commentsDelete&id=<?= $data['commentId'] ?>" class="text-dark" title="Pour supprimer le commentaire"><i class="p-2 fas fa-trash-alt fa-lg"></i></a>
+                <a href="index.php?action=commentsDelete&id=<?= $data['commentId'] ?>" onclick=" return confirm('Etes-vous sûr de vouloir supprimer ce commentaire ? ');" class="text-dark" title="Pour supprimer le commentaire"><i class="p-2 fas fa-trash-alt fa-lg"></i></a>
             </div>    
         </div>
        
         <?php
         }
     }
-$comments->closeCursor();
 
 $content = ob_get_clean(); 
-require('view/template.php'); 
+require('view/frontend/template.php'); 
 
 ?>
